threaded virtual machine
